feat(hearts): refactor into multi-directory
